+++ demo.py
@@ -51,10 +51,6 @@ for site in sites:
 
     # Start by visiting the page
     command_sequence.get(sleep=3, timeout=60)
-    command_sequence.save_screenshot()
-    command_sequence.screenshot_full_page()
-    command_sequence.dump_page_source()
-    command_sequence.recursive_dump_page_source()
 
     # Run commands across the three browsers (simple parallelization)
     manager.execute_command_sequence(command_sequence)
